Phase 1: address review nits and finalize hierarchy foundation
Agent-Logs-Url: https://github.com/starbounded-dev/LuxEngine/sessions/d53382d2-2538-494a-889e-b6b51fec073e

+++ Core/Source/Lux/Scene/Entity.cpp
@@ -18,7 +18,7 @@ namespace Lux {
 			return {};
 
 		const auto& relationship = m_Scene->m_Registry.get<RelationshipComponent>(m_EntityHandle);
-		if ((uint64_t)relationship.ParentHandle == 0)
+		if (relationship.ParentHandle == 0)
 			return {};
 
 		return m_Scene->GetEntityByUUID(relationship.ParentHandle);
